Add react routes & update App component
